backend/pyproject.toml
[project]
name = "crm-ai-cockpit-backend"
version = "0.1.0"
description = "FastAPI backend for CRM + AI Cockpit MVP"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.0",
  "asyncpg>=0.30.0",
  "cryptography>=43.0.0",
  "httpx>=0.27.0",
  "pgvector>=0.3.6",
  # passlib 1.7.4 is incompatible with bcrypt 4.1+ / 5.x (__about__ removed; hashing errors).
  "bcrypt>=4.0.1,<4.1.0",
  "email-validator>=2.0.0",
  "fastapi>=0.115.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.6.1",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.17",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.32.0",
]

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
